+++ src/ResourceManager/Billing/ChangeLog.md
@@ -19,6 +19,10 @@
 -->
 ## Current Release
 
+## Version 0.15.0
+* New Cmdlet Get-AzureRmEnrollmentAccount
+  - cmdlet to retrieve enrollment accounts
+
 ## Version 0.14.0
 * Add support for online help
     - Run Get-Help with the -Online parameter to open the online help in your default Internet browser
@@ -43,8 +47,8 @@
 * New Cmdlet Get-AzureRmBillingPeriod
     - cmdlet to retrieve azure billing periods of the subscription.
 * Update Cmdlet Get-AzureRmBillingInvoice
-	- new property BillingPeriodNames
-	- output in list view
+    - new property BillingPeriodNames
+    - output in list view
 
 ## Version 0.11.0